requirements.txt
asgiref==3.8.1
colorama==0.4.6
Django==4.0.3
django-cors-headers==4.4.0
django-filter==24.2
djangorestframework==3.13.1
djangorestframework-simplejwt==4.7.2
drf-yasg==1.21.7
inflection==0.5.1
iniconfig==2.0.0
lxml==5.2.2
packaging==24.1
pluggy==1.5.0
PyJWT==2.8.0
pytest==8.2.2
pytz==2024.1
PyYAML==6.0.1
setuptools==70.1.0
sqlparse==0.5.0
tzdata==2024.1
uritemplate==4.1.1